보상 트랜잭션 구현 #16

+++ member-service/src/main/java/com/memberservice/adapter/dto/ProductDto.java
@@ -1,5 +1,5 @@
 package com.memberservice.adapter.dto;
 
-public record ProductDto(Long productId, String productName, int price, String thumbnailUrl) {
+public record ProductDto(Long productId, String productName, Long price, String thumbnailUrl) {
 
 }
