+++ member-service/src/main/java/com/memberservice/controller/internal/res/CartDto.java
@@ -6,9 +6,9 @@ import com.memberservice.entity.Cart;
 public record CartDto(
     Long productId,
     String productName,
-    int price,
+    Long price,
     String thumbnailUrl,
-    int quantity
+    Long quantity
 ) {
 
     public CartDto(Cart cart) {
