+++ order-service/src/main/java/com/orderservice/adapter/req/ProcessDeliveryRequest.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 public record ProcessDeliveryRequest(
     Long productId,
     String productName,
-    int quantity,
+    Long quantity,
     Long orderLineId,
     Long deliveryAddressId) {
