+++ order-service/src/main/java/com/orderservice/adapter/req/ProcessPaymentRequest.java
@@ -7,8 +7,8 @@ public record ProcessPaymentRequest(
     Long memberId,
     Long orderLineId,
     Long paymentMethodId,
-    int totalPrice,
-    int discount
+    Long totalPrice,
+    Long discount
 ) {
 
 }
