+++ payment-service/src/main/java/com/paymentservice/usecase/dto/ProcessPaymentDto.java
@@ -7,8 +7,8 @@ public record ProcessPaymentDto(
     Long memberId,
     Long orderLineId,
     Long paymentMethodId,
-    int totalPrice,
-    int discount
+    Long totalPrice,
+    Long discount
 ) {
 
 }
