+++ product-service/src/main/java/com/productservice/controller/internal/res/ProductDto.java
@@ -3,7 +3,7 @@ package com.productservice.controller.internal.res;
 import com.productservice.entity.Product;
 import com.productservice.entity.ProductImage;
 
-public record ProductDto(Long productId, String productName, int price, String thumbnailUrl) {
+public record ProductDto(Long productId, String productName, Long price, String thumbnailUrl) {
     public ProductDto(Product product) {
         this(
             product.getId(),
