+++ product-service/src/main/java/com/productservice/usecase/dto/RegisterProductDto.java
@@ -9,8 +9,8 @@ import org.springframework.web.multipart.MultipartFile;
 public record RegisterProductDto(
     String name,
     String description,
-    int price,
-    int stock,
+    Long price,
+    Long stock,
     Set<String>tags,
     List<MultipartFile>productImages
 ) {
